Add separate module for nullspace functions
