+++ firedrake/slate/preconditioners/__init__.py
@@ -1,2 +1,2 @@
-from firedrake.slate.preconditioners.hybridization_mixed import *  # noqa: F401
+from firedrake.slate.preconditioners.hybridized_mixed import *  # noqa: F401
 from firedrake.slate.preconditioners.static_condensation import *  # noqa: F401
